requirements.txt
# shows python dependencies for the project

pandas
biopython
sqlalchemy
openpyxl
pymysql
